requirements.txt
beautifulsoup4==4.13.3
PyPDF2==3.0.1
Requests==2.32.3
selenium==4.29.0
pillow==11.1.0